Reorganize frontend directory structure and add local frontend
