+++ frontend/.gitignore
@@ -1,16 +1,14 @@
-# See https://help.github.com/ignore-files/ for more about ignoring files.
-
-src/**/*.css
-!src/styles/icons/style.css
+*.css
+!shared/styles/icons/style.css
 
 # dependencies
-/node_modules
+node_modules/
 
 # testing
-/coverage
+coverage/
 
 # production
-/build
+build/
 
 # misc
 .DS_Store
