+++ frontend/local/public/manifest.json
@@ -1,0 +1,15 @@
+{
+    "short_name": "IRYO",
+    "name": "IRYO cloud",
+    "icons": [
+        {
+            "src": "favicon.ico",
+            "sizes": "64x64 32x32 24x24 16x16",
+            "type": "image/x-icon"
+        }
+    ],
+    "start_url": "./index.html",
+    "display": "standalone",
+    "theme_color": "#000000",
+    "background_color": "#ffffff"
+}
