+++ frontend/local/src/containers/app/index.js
@@ -1,0 +1,39 @@
+import React from "react"
+import { Route, Link, NavLink } from "react-router-dom"
+import { connect } from "react-redux"
+import { bindActionCreators } from "redux"
+
+import Alert from "shared/containers/alert"
+import { close } from "shared/modules/alert"
+
+class App extends React.Component {
+    componentWillReceiveProps(nextProps) {
+        if (nextProps.location.pathname !== this.props.location.pathname) {
+            this.props.close()
+        }
+    }
+
+    logout() {
+        localStorage.removeItem("token")
+    }
+
+    render() {
+        return (
+            <div>
+                <main className="container">local</main>
+            </div>
+        )
+    }
+}
+
+const mapStateToProps = state => ({})
+
+const mapDispatchToProps = dispatch =>
+    bindActionCreators(
+        {
+            close
+        },
+        dispatch
+    )
+
+export default connect(mapStateToProps, mapDispatchToProps)(App)
