+++ frontend/package.json
@@ -1,0 +1,31 @@
+{
+    "private": true,
+    "workspaces": ["cloud", "local", "shared"],
+    "dependencies": {
+        "bootstrap": "^4.0.0",
+        "history": "^4.7.2",
+        "immer": "^0.8.5",
+        "jwt-decode": "^2.2.0",
+        "lodash": "^4.17.4",
+        "node-sass-chokidar": "^0.0.3",
+        "npm-run-all": "^4.1.2",
+        "react": "^16.2.0",
+        "react-dom": "^16.2.0",
+        "react-redux": "^5.0.6",
+        "react-router-dom": "^4.2.2",
+        "react-router-redux": "^5.0.0-alpha.9",
+        "react-scripts": "2.0.0-next.47d2d941",
+        "redux": "^3.7.2",
+        "redux-thunk": "^2.2.0",
+        "shared": "0.1.0"
+    },
+    "browserslist": {
+        "development": ["last 2 chrome versions", "last 2 firefox versions", "last 2 edge versions"],
+        "production": [">1%", "last 4 versions", "Firefox ESR", "not ie < 11"]
+    },
+    "prettier": {
+        "semi": false,
+        "tabWidth": 4,
+        "printWidth": 160
+    }
+}
