+++ frontend/shared/modules/authentication.js
