+++ frontend/shared/package.json
@@ -1,0 +1,4 @@
+{
+    "version": "0.1.0",
+    "name": "shared"
+}
