+++ frontend/shared/registerServiceWorker.js
