+++ frontend/shared/styles/fonts/SharpGroteskSmBold25/SharpGroteskSmBold25.woff2
