+++ frontend/shared/styles/icons/fonts/ElegantIcons.eot
