+++ frontend/shared/styles/icons/fonts/ElegantIcons.ttf
