+++ frontend/shared/styles/icons/fonts/ElegantIcons.woff
