+++ frontend/shared/styles/icons/style.css
